requirements.txt
Flask>=2.0.1
Flask-WTF>=1.0.0
Werkzeug>=2.0.1
Jinja2>=3.0.1
beautifulsoup4>=4.9.3
pdfkit>=1.0.0
requests>=2.31.0
gunicorn>=20.1.0
py_webauthn>=0.0.4
Flask-SQLAlchemy>=2.5.1
weasyprint>=52.5
python-dotenv>=0.19.0
Pillow>=10.0.0
WTForms>=3.0.0
email-validator>=1.1.3
Flask-Login>=0.5.0 